First draft of splitting out tasks_norm_lifecycle_template
